+++ src/alerts/alerts.constants.js
@@ -1,5 +1,0 @@
-const INITIAL_STATE = {
-  alerts: [],
-};
-
-export default INITIAL_STATE;