+++ src/cards/index.js
@@ -1,4 +1,0 @@
-export { CardContent } from './card-content.component.jsx';
-export { CardHeader } from './card-header.component.jsx';
-export { Card } from './card.component.jsx';
-export { Cards } from './cards.component.jsx';