+++ src/floating-select/index.js
@@ -1,2 +1,0 @@
-export FloatingSelect from './floating-select.component.jsx';
-export FloatingSelectCreatable from './floating-select-creatable.component.jsx';