+++ src/index.js
@@ -1,13 +1,15 @@
 import '../styles/main.scss';
 
-export * from './icons/index.js';
-export * from './alerts/index.js';
-export * from './cards/index.js';
+export * from '@opuscapita/react-alerts';
+export * from '@opuscapita/react-cards';
+export * from '@opuscapita/react-floating-select';
+export * from '@opuscapita/react-grid';
+export * from '@opuscapita/react-icons';
+
 export * from './confirm-dialog/index.js';
 export * from './dropdown-container/index.js';
 export * from './dropdown-menu/index.js';
 export * from './dropdown-multi-select/index.js';
-export * from './floating-select/index.js';
 export * from './responsive-navbar/index.js';
 export * from './searchbar/index.js';
 export * from './spinner/index.js';
@@ -16,5 +18,4 @@ export * from './layouts/application/index.js';
 export * from './extended-search/index.js';
 export * from './wizard/index.js';
 export * from './menu/index.js';
-export * from './datagrid/index.js';
 export * from './constants/index.js';